_updateChain in Sync
